Combine schema.go and mongo_driver.go into mongo.go file
